Moving NavigationBar and NavigationItem to /layout/. Removing old Navigation component. Updating import.
