+++ src/App.jsx
@@ -3,7 +3,7 @@ import { Outlet } from 'react-router-dom';
 import H1 from 'elements/H1';
 
 import AppHeader from 'layout/AppHeader';
-import NavigationBar from 'components/NavigationBar';
+import NavigationBar from 'layout/NavigationBar';
 
 import './output.css';
 import './App.css';
